requirements.txt
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
pillow>=8.3.0
tqdm>=4.62.0
tensorboard>=2.7.0
pyyaml>=5.4.0
matplotlib>=3.4.0